Add Computation.Parsing and .Console projects
(should be the last ones for now)

.Console will be used to figure out how the parsing needs to be
implemented and to have something to input stats to be parsed,
check for unknown stats, etc., once parsing is done.

.Parsing will use the matchers from .Data and combine their matches
into one result to be used by Computation. Most of the stuff
currently in Computation will probably be moved there.

Add IParser as the main interface for interacting with parsing
and a simple ToString-calling decorator as its first implementation.