Refactor IDamageStatBuilder.With(ItemSlot) to damage related
like the previous commit.
It now takes a new enum as argument that specifies the hand as it was only used with ItemSlot.MainHand and ItemSlot.OffHand.
To make up for it no longer being a condition, add condition version to IConditionBuilders. This one will also need to work with stats that are not IDamageRelatedStatBuilder.

Change all With methods in IDamageRelatedStatBuilder to return that type instead of IDamageStatBuilder (should have been changed in the previous commit).