Change modelling of "modifiers to ... apply to ..." stats
IStatBuilder.ApplyModifiersTo() is now in IDamageRelatedStatBuilder and the parameter is either a damage source or ailments. Also limited by forms now.
Implementing it as behaviors as was previously planned is not possible because it couldn't prevent modifiers from applying multiple times.
(for it to be implementable now, IStatBuilder must be able to build to different values for the different stats and the modifier's form must be passed to it)