Use Impale in DataDrivenMechanics
Implemented as a more multiplier to EffectiveDamageMultiplier for physical skill damage. The 10% stored damage is multiplied by Impale effect, stack count (if this is set to the maximum, the DPS represents the damage after ramping up, if set to 0, it is the first hit) and chance to Impale on hit.